Do not panic on unsupported compression type
An unsupported compression type can legitimately happen when we catch a
region in the middle of a server updating it. This error can be nicely
handled by simply retrying the read, but the panic disallows this.

Instead of panicking, simply introduce a new error type and return it
when we encounter such a case.